Cleaned up create_submission.py a bit
Made main's argument the parsed argument parser, fixed imports, and changed the catalyst way of inference to using `load_weights_infer`, adapted the documentation to the new changes, removed some comments. and adapted the thresholds to be 3500 pixels as the min size instead of 10,000. Also, I added checkpoint_path as an argparser argument, same as in create_classification_submission.py